Revert "updated more id data types"
This reverts commit 17b74259

+++ threatconnect-app/apps-core/src/main/java/com/threatconnect/app/apps/AppConfig.java
@@ -121,9 +121,9 @@ public abstract class AppConfig
 		return getString(TC_API_PATH);
 	}
 	
-	public Long getTcUserId()
+	public Integer getTcUserId()
 	{
-		return getLong(TC_USER_ID);
+		return getInteger(TC_USER_ID);
 	}
 	
 	public Integer getTcSpaceElementId()
